% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjWgt.r
\name{AdjWgt}
\alias{AdjWgt}
\title{Adjust design weights for nonsampling error.}
\usage{
AdjWgt(dat, popn, evalstatus = "EvalStatus", wgt = "wgt")
}
\arguments{
\item{dat}{Data frame containing columns at least for \code{Site},
\code{WYear}, \code{Year}, the outcome of interest, and the survey
evaluation status \code{EvalStatus}. See section "Options for variable
\code{EvalStatus}" below.}

\item{popn}{Takes the value of \code{"Target"} or \code{"Sampled"} to reflect
inference to either the original target population or to the subpopulation
of sites that would not be subject to nonresponse, respectively.}

\item{evalstatus}{Status of the survey at each site relative to frame error
and nonresponse error. Note that this field is calculated separately for
each year at a site.}

\item{wgt}{Original design weight. The design weights for a given year should
sum to the size of the population before accounting for any frame error.}
}
\value{
Returns a vector of adjusted design weights that should sum to the
  extent of the population of interest after accounting for frame error
  (\code{popn = "Target"}) or frame and nonresponse error (\code{popn =
  "Sampled"}).
}
\description{
Using the \code{evalstatus} field in the design file, adjust the
  design weights for frame error and nonsampling error.
}
\details{
Adjusts the original sampling design weights for frame and
  nonresponse error. See (Starcevich et al. 2016) for more information on
  weighting adjustments and populations of inference. See (Starcevich et al.
  2017) for an example on applying the AdjWgt_FrameNR function to a
  monitoring data set.
}
\section{Options for variable \code{evalstatus}}{

  
  The \code{evalstatus} field requires specific formatting for weighting
  adjustment.  Valid options include:
  
  \tabular{ll}{
  
  \code{"Target - Surveyed"} \tab Site is in target population and was
  successfully surveyed. \cr \code{"Target - Not Surveyed"}     \tab Site is
  in target population but was not successfully surveyed.  \cr 
  \code{"Non-Target"}    \tab Site is not in target population, not surveyed.
  \cr \code{"Not Evaluated"}     \tab Site was not evaluated as target or
  not, not surveyed. \cr
  
  }
}
\examples{
\dontrun{
#  ---- Read example data set.
	LakeData$AdjWgt[LakeData$WYear == 0] = AdjWgt(dat=LakeData[LakeData$WYear == 0,],
	                                              popn="Target",
	                                              evalstatus = "EvalStatus",
	                                              wgt = "wgt")
}


}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\references{
Starcevich, L. A., G. DiDonato, T. McDonald, and J. Mitchell.
  (2016). A GRTS user's manual for the SDrawNPS package: A graphical user
  interface for generalized random tessellation stratified (GRTS) sampling 
  and estimation. Natural Resource Report NPS/PWRO/NRR-2016/1233. National
  Park Service, Fort Collins, Colorado.
  
  Starcevich, L. A., T. McDonald, A. Chung-MacCoubrey, A. Heard, and J.
  Nesmith. (2017). Trend Estimation for Complex Survey Designs. Natural
  Resource Report NPS/xxxx/NRR-2017/xxxx. National Park Service, Fort
  Collins, Colorado.
}

