% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendNPS.r
\docType{data}
\name{Cover_Acro}
\alias{Cover_Acro}
\title{\emph{Acrosiphonia} binary cover data.}
\format{A data frame with 1794 rows and 10 variables.

\describe{The file contains binary indicators of \emph{Acrosiphonia} cover
collected between 2009 and 2015 across two parks, Katmai National Park (KATM)
and Kenia Fjords National Park (KEFJ) in the Southwest Alaskan Network. Data
hierarchy includes Elevation class within Quadrat within Site.  The variables
are as follows:

\itemize{ 
  \item Site. Sampling unit selected with GRTS sampling. 
  \item Park. National Park. \code{KATM} = Katmai NP and \code{KEFJ} = Kenai Fjords NP.
  \item Quad. Quadrat within Site. 
  \item Elev. Elevation class within \code{Quad} Quadrat.
  \item Year. Survey year formatted as a factor for random effects estimation.
  \item WYear. Scalar year covariate for trend analysis.
  \item Gradient. Site gradient (degrees).
  \item Long. Site longitude.   
  \item Lat. Site latitude.
  \item wgt. Design weight from the sample draw.
  \item PanelWt. Weight to account for temporal revisit design given sample inclusion.
}
}}
\usage{
data(Cover_Acro)
}
\description{
\emph{Acrosiphonia} binary cover data.
}
\references{
Alaska Ocean Observing System Gulf of Alaska Data Integration
  Portal. 2017. Nearshore: Intertidal Systems in Gulf of Alaska data
  available at: 
  \url{http://portal.aoos.org/gulf-of-alaska.php#metadata/53c052b6-8874-46d1-b40a-acc615a3879a/project/files}.
}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\keyword{datasets}
