% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{"Expit" transform, or the backwards logit transform.}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Apply the standard backwards logit transform to a numeric vector
  of values.
}
\details{
The \code{expit} function simply applies the standard backwards
  logit transform to the provided numeric vector \code{x}: \eqn{1 / (1 +
  exp(-x))}.
}
\examples{
x <- c(-0.8472979,-1.3862944,2.1972246)
y <- expit(x)

y
# 0.3 0.2 0.9
}
\seealso{
\code{TrendNPS::logit}
}
\author{
Leigh Ann Starcevich and Jason Mitchell of Western EcoSystems 
  Technology, Inc.
}
