% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearizationVar.r
\name{LinearizationVar}
\alias{LinearizationVar}
\title{Calculate linearization variance of trend (Skinner, et al. 1989)}
\usage{
LinearizationVar(Site, wij, xij, eij, varYij)
}
\arguments{
\item{Site}{vector of PSUs.}

\item{wij}{inclusion weight = PSU wt * SSU wt.}

\item{xij}{year var for trend estimation.}

\item{eij}{residual.}

\item{varYij}{total variance for yij, estimated from PWIGLS estimates of
variance from P&O.}
}
\value{
Something cool.
}
\description{
Calculate the linearization variance of the trend coefficient
}
\details{
Something cool.
}
\examples{
\dontrun{
#  ---- Read example data set.
SEKIANC_orig = LinearizationVar(Site,wij,xij,eij,varYij)
}


}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\references{
Skinner 1989., p. 82-83.
}
\seealso{
\code{TrendNPS_Cont}, \code{LinearizationVar}
}

