% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearizationVar_StRS.r
\name{LinearizationVar_StRS}
\alias{LinearizationVar_StRS}
\title{Calculate linearization variance of trend for stratified sampling.}
\usage{
LinearizationVar_StRS(Site, wij, xij, eij, varYij, str1prop)
}
\arguments{
\item{Site}{vector of Site names.}

\item{wij}{vector of inclusion weights = Site-level design weight * Year-level panel weight.}

\item{xij}{vector of shifted year variables (WYear) for trend estimation.}

\item{eij}{vector of residuals from PWIGLS trend model.}

\item{varYij}{vector of total variance estimates for yij, estimated from PWIGLS variance 
components estimates.}

\item{str1prop}{Proportion of the population represented by the first level of the 
stratication variable.}
}
\value{
Scalar variance estimate of the trend regression coefficient.
}
\description{
Calculate the linearization variance of the trend coefficient with a Taylor
series approximation when two-level stratification is used and a separate-slopes trend 
model is applied. This is an internal function called from PWIGLS_ALL.
}
\details{
The linearization variance is based on a Taylor series approximation. See 
Skinner (1989, p. 82-83) for more information.
}
\examples{
\dontrun{
#  ---- Read example data set.
TrendVar = LinearizationVar_StRS(Site,wij,xij,eij,varYij,str1prop)
}


}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\references{
Skinner, C. J., D. Holt, and T. M. F. Smith. 1989. Analysis of Complex Surveys. 
New York: Wiley.
}
\seealso{
\code{TrendNPS_Cont}, \code{LinearizationVar}
}

