% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit transform.}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Apply the standard logit transform to a numeric vector of
  values.
}
\details{
The \code{logit} function simply applies the standard logit
  transform to the provided numeric vector \code{x}: \eqn{log(x / (1 - x))}.
}
\examples{
x <- c(0.3,0.2,0.9)
y <- logit(x)

y
# -0.8472979 -1.3862944  2.1972246
}
\seealso{
\code{TrendNPS::expit}
}
\author{
Leigh Ann Starcevich and Jason Mitchell of Western EcoSystems 
  Technology, Inc.
}
