% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWIGLS_ALL.r
\name{PWIGLS_ALL}
\alias{PWIGLS_ALL}
\title{Calculate cool stuff.}
\usage{
PWIGLS_ALL(Z, dat, slope, type, stratum, stage1wt, stage2wt)
}
\arguments{
\item{Z}{Something.}

\item{dat}{Data frame containing columns at least for \code{Site}, 
\code{WYear}, \code{Year}, and the continuous outcome of interest \code{Y}.
See section "Data frame \code{dat}" below.}

\item{slope}{Logical value indicating inclusion of a random site-level slope
effect in the variance components structure used for the \code{"PO"} and 
\code{"PWIGLS"} trend methods.  Site- and year-level random intercept terms
included as default.}

\item{type}{Scaling type when \code{method="PWIGLS"}. Valid values include 
\code{"Aonly"}, \code{"A"}, \code{"AI"}, \code{"B"}, \code{"BI"} 
\code{"C"}. See section "Options for variable \code{type}" below.}

\item{stratum}{Text string identifying an optional two-level stratification 
factor in \code{dat}.}

\item{stage1wt}{Design weights from the original sample draw without 
accounting for temporal revisit designs.}

\item{stage2wt}{Panel inclusion weights for each site each year.}
}
\value{
Something cool.
}
\description{
Calculate cool stuff.
}
\details{
Something cool.
}
\section{Options for variable \code{type}}{

  
  Selection of \code{method="PWIGLS"} requires further specification of 
  argument \code{type}.  Valid options include
  
  \tabular{ll}{
  
  \code{"Aonly"} \tab probability weighting but no scaling at either stage \cr
  \code{"A"}     \tab scaling for panel weights with the mean site-level design weight  \cr
  \code{"AI"}    \tab scaling for panel weights with the mean site-level design weight, but no site-level scaling  \cr
  \code{"B"}     \tab scaling for panel weights with the effective mean site-level design weight  \cr
  \code{"BI"}    \tab scaling for panel weights with the effective mean site-level design weight, but no site-level scaling  \cr
  \code{"C"}     \tab scaling only at the year level with the inverse of the average year-level weight \cr
  
  }
}
\examples{
\dontrun{
#  ---- Read example data set.
SEKIANC_orig = LinearizationVar(Site,wij,xij,eij,varYij)
}


}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\references{
Asparouhov, T. (2006).  General multi-level modeling with 
  sampling weights. Communications in Statistics - Theory and Methods 35: 
  439-460.

  Pfeffermann, D., C.J. Skinner, D.J. Holmes, H. Goldstein, and J. Rasbash 
  (1998).  Weighting for unequal selection probabilities in multilevel 
  models. Journal of the Royal Statistical Society, Series B 60(1): 23-40.
}
\seealso{
\code{TrendNPS_Cont}, \code{LinearizationVar}
}

