% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendNPS_Binary.R
\name{TrendNPS_Binary}
\alias{TrendNPS_Binary}
\title{Trend for Binary Outcomes from Complex Survey Designs}
\usage{
TrendNPS_Binary(alpha, dat, method, slope = TRUE, type = NA, stratum = NA,
  Y, lat = NA, long = NA, stage1wt = NA, stage2wt = NA, str1prop = NA,
  nbhd = TRUE)
}
\arguments{
\item{alpha}{The probability of a Type I error.}

\item{dat}{Data frame containing columns at least for \code{Site}, 
\code{WYear}, \code{Year}, and the binary outcome of interest \code{Y}.
See section "Data frame \code{dat}" below.}

\item{method}{Method for trend estimation entered as a string. Valid values 
include \code{"PO"} for the generalized linear mixed model extension of the 
Piepho and Ogutu (2002) unreplicated linear mixed model, \code{"SLRDB"} for 
simple linear regression of design-based estimates, \code{"WLRDB"} for 
weighted linear regression of design-based estimates, or \code{"PWIGLS"} for 
probability-weighted iterative generalized least squares (Pfeffermann et al. 
1998, Asparouhov 2002).}

\item{slope}{Logical value indicating inclusion of a random site-level slope
effect in the variance components structure used for the \code{"PO"} and 
\code{"PWIGLS"} trend methods.  Site- and year-level random intercept terms
included as default.}

\item{type}{Scaling type when \code{method="PWIGLS"}. Valid values include 
\code{"Aonly"}, \code{"A"}, \code{"AI"}, \code{"B"}, \code{"BI"} 
\code{"C"}. See section "Options for variable \code{type}" below.}

\item{stratum}{Text string identifying an optional two-level stratification 
factor in \code{dat}.}

\item{Y}{Text string indicating the binary outcome variable in \code{dat}.}

\item{lat}{Site latitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{long}{Site longitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{stage1wt}{Design weights from the original sample draw without 
accounting for temporal revisit designs.}

\item{stage2wt}{Panel inclusion weights for each site each year.}

\item{str1prop}{Proportion of the first stratum in the population in a 
two-level stratification variable.}

\item{nbhd}{Logical indictor for using the neighborhood variance estimator 
when the \code{method="WLRDB"}.  If \code{FALSE}, \code{"WLRDB"} calculates
the design-based mean estimate's standard error assuming independent random
sampling.}
}
\value{
A list containing three or four elements, depending on the specified trend method. 
  The first element of the list contains a data frame containing the estimated intercept,  
  trend estimate, trend standard error, estimated variance components, and degrees of freedom  
  used for trend testing and confidence intervals construction.
  
  The contents of the first list element, ModelEstimates, are:
  
  \tabular{rl}{
  
  \code{mu}      \tab	Estimated intercept from the trend model/ \cr
  \code{trend}   \tab	Estimated trend of the outcome \code{Y} on the link function scale from the trend model.  \cr
  \code{SEtrend} \tab	Estimated standard error of the trend estimate. \cr
  \code{sig2a}   \tab	Estimated site-to-site variation. \cr
  \code{sig2b}   \tab	Estimated year-to-year variation. \cr
  \code{sig2t}   \tab	Estimated variation among site-level slopes. \cr
  \code{sig2e}   \tab	Estimated residual variation. \cr
  \code{eta}     \tab	Degrees of freedom used to test for trend and calculate confidence intervals. \cr
  
  }
  
  The contents of the second list element, TrendTest, include the trend coefficient on the link scale, 
  the trend standard error, the t-statistic, degrees of freedom, and p-value for a two-sided test of no trend.
  
  The contents of the third list element, TrendCI, include the back-transformed trend on the original scale 
  and the lower and upper (1-\code{alpha})*100\%-confidence interval bounds. 
  
  A fourth list element, DBests, contains the annual design-based estimates and is returned only when the 
  SLRDB and WLRDB trend approaches are used.  The design-based estimates may be used to assess the assumption
  of independence for the linear regression models.  The output of the DBests list element include:
  
  \tabular{rl}{
  
  \code{Year}      \tab	Year of survey. \cr
  \code{Est.Mean}   \tab	Design-based estimate of the mean of the outcome of interest. \cr
  \code{SE} \tab	   Estimated standard error of design-based mean estimate via neighborhood 
  variace estimator if \code{nbhd = TRUE}. \cr
  \code{WYear}   \tab	WYear variable. \cr
  \code{Resid}     \tab	Residuals from the linear model fit. \cr
  
  }
}
\description{
\code{TrendNPS_Cont} fits trend models of binary outcomes 
  
  \enumerate{ \item the unreplicated generalized linear mixed model with 
  variance components as specified by Piepho and Ogutu (2002) that does 
  not incorporate design weights ("PO");
  
  \item simple linear regression of annual design-based estimates ("SLRDB");
  
  \item weighted linear regression of annual design-based estimates (WLRDB);
  
  \item and probability-weighted iterative generalized least squares (PWIGLS) 
  
  } Fixed effects structure includes a year term for trend estimation and an 
  optional two-level stratification factor. The function assumes random site 
  and year intercepts, and an optional random site-level slope effect may be 
  included.
}
\details{
For long-term monitoring, \code{TrendNPS_Binary} assumes a single annual sampling 
  occasion per unique record of \code{WYear} and \code{Site} in data frame \code{dat}.  
  
  Sampling (\code{stage1wt}) and temporal-revisit (\code{stage2wt}) design 
  weights must be adjusted for any missing data or frame error prior to trend
  analysis. See Oh and Scheuren (1983), Little and Rubin (2002), and 
  Starcevich et al. (2016).
  
  For background on the variance structure, see Piepho & Ogutu (2002).  For information 
  on the probability-weighted generalized least squares approach, see Pfeffermann 
  et al. (1998), Asparouhov (2006), and Starcevich et al. (2017).
}
\section{Method requirements}{

  
  Spatially balanced samples utilizing \code{method="SLRDB"} or 
  \code{method="WLRDB"} require specification of \code{lat} and \code{long} 
  on setting the neighborhood variance estimator \code{nbhd=TRUE}.  If 
  spatial balance cannot be assumed, set option \code{nbhd=FALSE} for 
  non-spatially balanced designs so that the \code{"SLRDB"} and 
  \code{"WLRDB"} methods assume independent random sampling for standard 
  errors.
  
  The stratification variable \code{stratum} is required for all methods when
  design strata are used.  Note that \code{method="PWIGLS"} further requires
  a value for \code{type}.  See section "Options for variable \code{type}"
  below.
  
  Variables \code{stage1wt} and \code{stage2wt} must be specified for all 
  methods except the unweighted \code{method="PO"}.
}

\section{Data frame \code{dat}}{

  
  Data frame \code{dat} requires at least variables \code{Site}, 
  \code{WYear}, \code{Year}, and \code{Y}.  Variable \code{Site} defines the 
  sampling unit.
  
  Underlying statistical functions called by \code{TrendNPS_Cont} require two
  separate temporal variables: \code{Year} is a factor and \code{WYear} is a 
  scalar-year value.  \code{WYear} is shifted so that the year for which \code{Y}
  is least variable is centered on 0. See Piepho & Ogutu (2002) for more 
  information.
  
  Ultimately, \code{WYear} represents time in the fixed-effects portion of the  
  trend model. The regression coefficient associated with \code{WYear} provides
  the basis for trend estimation and testing. 
  
  Variable \code{Year} enters the model as a random effect and provides 
  the basis for estimation of random year-to-year variation among survey 
  occasions.
  
  Variable \code{Y} must contain only binary values coded as 0 or 1.
}

\section{Options for variable \code{type}}{

  
  Selection of \code{method="PWIGLS"} requires further specification of 
  argument \code{type}.  Valid options include
  
  \tabular{rl}{
  
  \code{"Aonly"} \tab Probability weighting but no scaling at either stage. \cr
  \code{"A"}     \tab Scaling for panel weights with the mean site-level design weight.  \cr
  \code{"AI"}    \tab Scaling for panel weights with the mean site-level design weight, but no site-level scaling.  \cr
  \code{"B"}     \tab Scaling for panel weights with the effective mean site-level design weight.  \cr
  \code{"BI"}    \tab Scaling for panel weights with the effective mean site-level design weight, but no site-level scaling.  \cr
  \code{"C"}     \tab Scaling only at the year level with the inverse of the average year-level weight. \cr
  
  }
}

\examples{

#  ---- Read example data set.
data(Cover_Acro)

#  ---- Load dependent packages. 
pkgList <- c("lme4","lmerTest","spsurvey")
inst <- pkgList \%in\% installed.packages()
if (length(pkgList[!inst]) > 0) install.packages(pkgList[!inst])
lapply(pkgList, library, character.only = TRUE)

###########################
# Example 1: Trend analysis of a binary indicator of Acrosophonia presence
# with the PO approach: stratification and full random effects model.
TrendAcro_PO_StRS = TrendNPS_Binary(alpha=0.1,
dat=Cover_Acro,method="PO",slope=TRUE,type=NA,stratum="Park",Y="Y",
stage1wt="wgt",stage2wt="PanelWt",str1prop=0.13227) 

# $ModelEstimates
#           mu    trend    SEtrend     sig2a      sig2t       sigat     sig2b
#   -0.2992415 0.221928 0.09341594 0.2867116 0.01394005 -0.04732509 0.2411157

# $TrendTest
#      trend    SEtrend   z-stat    pvalue
#   0.221928 0.09341594 2.375698 0.0175158

# $TrendCIofOddsRatio
#   Annual Pct Change     CI low   CI high
#           0.2484815 0.07065702 0.4558408


# Example 2: Trend analysis of a binary indicator of Acrosophonia presence
# with the SLRDB approach for stratification.
TrendAcro_SLRDB_StRS = TrendNPS_Binary(alpha=0.1,
dat=Cover_Acro,method="SLRDB",slope=TRUE,type=NA,stratum="Park",Y="Y",
lat="Lat",long="Long",stage1wt="wgt",stage2wt="PanelWt",
str1prop=0.13227) 

# TrendAcro_SLRDB_StRS

# $ModelEstimates
#           mu    trend    SEtrend sig2a sig2t sigat sig2b
#   -0.9909923 0.211127 0.08966504     0     0     0     0
# 
# $TrendTest
#      trend    SEtrend   z-stat     pvalue
#   0.211127 0.08966504 2.354619 0.01854169
# 
# $TrendCIofOddsRatio
#  Annual Pct Change     CI low   CI high
#           0.2350692 0.06570992 0.4313426
# 
# $DBests
#   Year  Est.Mean         SE WYear      Resid
#   2008 0.1604049 0.06934590     0 -0.6642261
#   2009 0.2652986 0.05963881     1 -0.2387430
#   2010 0.4121248 0.10439841     2  0.2135496
#   2011 0.6416667 0.07341420     3  0.9402165
#   2012 0.5439923 0.11698602     4  0.3229097
#   2013 0.5875758 0.03719284     5  0.2893100
#   2014 0.4954618 0.06389670     6 -0.2939231
#   2015 0.4794624 0.04936033     7 -0.5690937

# Plot trend on log-odds scale
num <- TrendAcro_SLRDB_StRS$DBests$Est.Mean
denom <- 1-TrendAcro_SLRDB_StRS$DBests$Est.Mean#' 
TrendAcro_SLRDB_StRS$DBests$LogOdds = num / denom
plot(TrendAcro_SLRDB_StRS$DBests$Year,
     TrendAcro_SLRDB_StRS$DBests$LogOdds,
     xlab="Year", ylab="Odd ratio of cover proportion")
lines(2008:2015, exp(-0.9909923 + 0.211127*(0:7)), col=2)

# Plot trend on proportional cover scale
expit <- function(x){
  y <- 1 / (1 + exp(-x))
  return(y)
}
plot(TrendAcro_SLRDB_StRS$DBests$Year,
     TrendAcro_SLRDB_StRS$DBests$Est.Mean,
     xlab="Year", ylab="Cover proportion")
lines(2008:2015, expit(-0.9909923 + 0.211127*(0:7)), col=2)
}
\references{
Asparouhov, T. (2006).  General multi-level modeling with 
  sampling weights. Communications in Statistics - Theory and Methods 35: 
  439-460.
  
  Little, J. A. R., and D. B. Rubin. (2002). Statistical analysis with missing 
  data, 2nd edition. John Wiley and Sons, Inc., New Jersey.
  
  Oh, H. L., and F. J. Scheuren. (1983). Weighting adjustment for unit 
  nonresponse. Pages 143-184 in W.G. Madow, I. Olkin, and D.B. Rubin, 
  editors. Incomplete data and in sample surveys. (Vol. 2). Academic Press, 
  New York.
  
  Pfeffermann, D., C. J. Skinner, D. J. Holmes, H. Goldstein, and J. Rasbash 
  (1998).  Weighting for unequal selection probabilities in multilevel 
  models. Journal of the Royal Statistical Society, Series B 60(1): 23-40.
  
  Piepho, H. P., & Ogutu, J. O. (2002).  A simple mixed model for trend 
  analysis in wildlife populations.  Journal of Agricultural, Biological, and
  Environmental Statistics, 7(3), 350-360.
  
  Starcevich, L. A., G. DiDonato, T. McDonald, and J. Mitchell. (2016). A GRTS 
  user's manual for the SDrawNPS package: A graphical user interface for 
  generalized random tessellation stratified (GRTS) sampling and estimation. 
  Natural Resource Report NPS/PWRO/NRR-2016/1233. National Park Service, Fort
  Collins, Colorado.
  
  Starcevich, L. A., T. McDonald, A. Chung-MacCoubrey, A. Heard, and J. Nesmith.
  (2017). Trend Estimation for Complex Survey Designs. Natural Resource 
  Report NPS/xxxx/NRR-2017/xxxx. National Park Service, Fort Collins, 
  Colorado.
  R. Wolfinger. (1993). Laplace's approximation for nonlinear mixed models. 
  Biometrika 80(4): 791-795.
}
\seealso{
\code{lme4}, \code{lmerTest}, \code{spsurvey}
}
\author{
Leigh Ann Starcevich and Jason Mitchell of Western EcoSystems 
  Technology, Inc.
}
