% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendNPS_Cont.r
\name{TrendNPS_Cont}
\alias{TrendNPS_Cont}
\title{Trend for Continuous Outcomes from Complex Survey Designs}
\usage{
TrendNPS_Cont(dat, method, slope = TRUE, type = NA, stratum = NA, Y,
  lat = NA, long = NA, stage1wt = NA, stage2wt = NA, str1prop = NA,
  nbhd = TRUE)
}
\arguments{
\item{dat}{Data frame containing columns at least for \code{Site}, 
\code{WYear}, \code{Year}, and the continuous outcome of interest \code{Y}.
See section "Data frame \code{dat}" below.}

\item{method}{Method for trend estimation entered as a string. Valid values 
include \code{"PO"} for the Piepho and Ogutu (2002) unreplicated linear 
mixed model, \code{"SLRDB"} for simple linear regression of design-based 
estimates, \code{"WLRDB"} for weighted linear regression of design-based 
estimates, or \code{"PWIGLS"} for probability-weighted iterative 
generalized least squares (Pfeffermann et al. 1998, Asparouhov 2002).}

\item{slope}{Logical value indicating inclusion of a random site-level slope
effect in the variance components structure used for the \code{"PO"} and 
\code{"PWIGLS"} trend methods.  Site- and year-level random intercept terms
included as default.}

\item{type}{Scaling type when \code{method="PWIGLS"}. Valid values include 
\code{"Aonly"}, \code{"A"}, \code{"AI"}, \code{"B"}, \code{"BI"} 
\code{"C"}. See section "Options for variable \code{type}" below.}

\item{stratum}{Text string identifying an optional two-level stratification 
factor in \code{dat}.}

\item{Y}{Text string indicating the outcome variable in \code{dat}.}

\item{lat}{Site latitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{long}{Site longitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{stage1wt}{Design weights from the original sample draw without 
accounting for temporal revisit designs.}

\item{stage2wt}{Panel inclusion weights for each site each year.}

\item{str1prop}{Proportion of the first stratum in the population in a 
two-level stratification variable.}

\item{nbhd}{Logical indictor for using the neighborhood variance estimator 
when the \code{method="WLRDB"}.  If \code{FALSE}, \code{"WLRDB"} calculates
the design-based mean estimate's standard error assuming independent random
sampling.}
}
\value{
A displayed vector containing the trend estimate, along with its standard
  error, estimated variance components, and degrees of freedom used to test
  for trend and calculate confidence intervals.
  
  Additionally, a data frame containing each of 
  
  \tabular{rl}{
  
  \code{mu}      \tab	Estimated intercept from the trend model/ \cr
  \code{trend}   \tab	Estimated trend of the logged outcome \code{Y} from the trend model.  \cr
  \code{SEtrend} \tab	Estimated standard error of the trend estimate for the logged outcome \code{Y}. \cr
  \code{sig2a}   \tab	Estimated site-to-site variation. \cr
  \code{sig2b}   \tab	Estimated year-to-year variation. \cr
  \code{sig2t}   \tab	Estimated variation among site-level slopes. \cr
  \code{sig2e}   \tab	Estimated residual variation. \cr
  \code{eta}     \tab	Degrees of freedom used to test for trend and calculate confidence intervals. \cr
  
  }
}
\description{
\code{TrendNPS_Cont} fits trend models of continuous outcomes 
  for four approaches:
  
  \enumerate{ \item the unreplicated linear mixed model as specified by 
  Piepho and Ogutu (2002) that does not incorporate design weights ("PO");
  
  \item simple linear regression of annual design-based estimates ("SLRDB");
  
  \item weighted linear regression of annual design-based estimates (WLRDB);
  
  \item and probability-weighted iterative generalized least squares of a 
  linear mixed model (PWIGLS) with 6 types of scaling.
  
  } Fixed effects structure includes a year term for trend estimation and an 
  optional two-level stratification factor. The function assumes random site 
  and year intercepts, and an optional random site-level slope effect may be 
  included.
}
\details{
For long-term monitoring, \code{TrendNPS_Cont} assumes sampling 
  occasions occur no more than once per unique record of \code{WYear} and 
  \code{Site} in data frame \code{dat}.  [Tell people what to do to get 
  around this?]
  
  Sampling (\code{stage1wt}) and temporal-revisit (\code{stage2wt}) design 
  weights must be adjusted for any missing data or frame error prior to trend
  analysis. See Oh and Scheuren (1983), Little and Rubin (2002), and 
  Starcevich et al. (2016).
  
  For background on the unweighted linear mixed model, see Piepho & Ogutu
  (2002).  For information on the probability-weighted generalized least
  squares approach, see Pfeffermann et al. (1998), Asparouhov (2006), and
  Starcevich et al. (2017).
}
\section{Method requirements}{

  
  Spatially balanced samples utilizing \code{method="SLRDB"} or 
  \code{method="WLRDB"} require specification of \code{lat} and \code{long} 
  on setting the neighborhood variance estimator \code{nbhd=TRUE}.  Set 
  option \code{nbhd=FALSE} for non-spatially balanced designs so that the 
  \code{"WLRDB"} method [and \code{"SLRDB"}?] assumes independent random 
  sampling for standard errors.  These methods also require specification of 
  \code{stratum}.
  
  The stratification variable \code{stratum} is required for 
  \code{method="PO"} and \code{method="PWIGLS"}.  Note that 
  \code{method="PWIGLS"} further requires a value for \code{type}.  See 
  section "Options for variable \code{type}" below.
  
  Variables \code{state1wt} and \code{stage2wt} must be specified when 
  \code{method="PWIGLS"}.
}

\section{Data frame \code{dat}}{

  
  Data frame \code{dat} requires at least variables \code{Site}, 
  \code{WYear}, \code{Year}, and \code{Y}.  Variable \code{Site} serves as a 
  statistical-unit text descriptior.
  
  Underlying statistical functions called by \code{TrendNPS_Cont} require two
  separate temporal variables: \code{WYear} is a scalar-year value, while
  \code{Year} is a factor.
  
  Ultimately, \code{WYear} enters the model as a trend effect, and designates
  the survey occasions used in the fixed-effects portion of the trend model. 
  The regression coefficient associated with \code{WYear} provides the basis 
  for trend estimation and testing. In line with Piepho & Ogutu (2002), 
  function \code{TrendNPS_Cont} reassigns the recorded temporal unit 
  \code{WYear} in which the outcome \code{Y} is least variable as zero, with 
  all other values \code{WYear} calculated with respect to this redefined 
  zero point.
  
  Variable \code{Year} enters the model as a random effect, and identifies 
  unique values for each survey occasion.  As a random effect, it provides 
  the basis for estimation of random year-to-year variation among survey 
  occasions.
  
  Variable \code{Y} must contain only positive non-zero values.  Analyses 
  utilizing \code{method="PO"} or \code{method="PWIGLS"} assumes variable 
  \code{Y} has been previously transformed to a logarithmic scale.
}

\section{Options for variable \code{type}}{

  
  Selection of \code{method="PWIGLS"} requires further specification of 
  argument \code{type}.  Valid options include
  
  \tabular{rl}{
  
  \code{"Aonly"} \tab Probability weighting but no scaling at either stage. \cr
  \code{"A"}     \tab Scaling for panel weights with the mean site-level design weight.  \cr
  \code{"AI"}    \tab Scaling for panel weights with the mean site-level design weight, but no site-level scaling.  \cr
  \code{"B"}     \tab Scaling for panel weights with the effective mean site-level design weight.  \cr
  \code{"BI"}    \tab Scaling for panel weights with the effective mean site-level design weight, but no site-level scaling.  \cr
  \code{"C"}     \tab Scaling only at the year level with the inverse of the average year-level weight. \cr
  
  }
}
\examples{

#  ---- Read example data set.
data(SEKIANC)

#  ---- Load dependent packages. 
pkgList <- c("lme4","lmerTest","spsurvey")
inst <- pkgList \%in\% installed.packages()
if (length(pkgList[!inst]) > 0) install.packages(pkgList[!inst])
lapply(pkgList, library, character.only = TRUE)

###########################
#  ---- A1. PO method: full random effects model.
PO_ests <- TrendNPS_Cont(dat=SEKIANC, method="PO", type=NA, slope=TRUE, 
                         stratum=NA, Y="Y", lat="Lat", long="Long",
                         stage1wt="wgt", stage2wt="Panelwgt", str1prop=NA)
# Error: number of observations (=80) <= number of random effects (=92) for 
# term (1 + WYear | Site); the random-effects parameters and the residual 
# variance (or scale parameter) are probably unidentifiable

#  ---- Conclusion: the random effects are not estimable due to too few
#  ---- observations.  Reduce the random effects model by removing the 
#  ---- random slope term; i.e., set slope=FALSE.

#  ---- A2. PO method: reduced random effects model.
PO_ests <- TrendNPS_Cont(dat=SEKIANC, method="PO", type=NA, slope=FALSE,
                         stratum=NA, Y="Y", lat="Lat", long="Long",
                         stage1wt="wgt", stage2wt="Panelwgt", str1prop=NA)
round(PO_ests,4)

#     mu  trend SEtrend  sig2a sig2t sigat sig2b  sig2e    eta
#  3.964 0.0194  0.0322 0.3422     0     0     0 0.0161 3.6202

#  ---- Construct a 95\%-confidence interval on the trend estimate of 
#  ---- the logged outcome.
CI <- c(PO_ests$trend - qt(1-(0.05/2), PO_ests$eta) * PO_ests$SEtrend, 
        PO_ests$trend + qt(1-(0.05/2), PO_ests$eta) * PO_ests$SEtrend)
round(c(PO_ests$trend, CI),4)

# 0.0194 -0.0737  0.1125

#  ---- Construct a 95\%-confidence interval on the trend estimate of 
#  ---- the original outcome.  
round(exp(c(PO_ests$trend, CI)),4)

# 1.0196 0.9290 1.1191

#  ---- Conduct a two-sided test of trend for an alternative hypothesis 
#  ---- of trend in either direction.
t_stat <- PO_ests$trend/ PO_ests$SEtrend
p_value <- 2*(1-pt(abs(t_stat), PO_ests$eta, lower=TRUE))
round(data.frame(t_stat,p_value) ,4)

# 0.6038 0.5817


###########################
#  ---- B1. WLRDB method.
WLRDB_ests <- TrendNPS_Cont(dat=SEKIANC, method="WLRDB", type=NA, 
                            slope=FALSE, stratum=NA, Y="Y", lat="Lat",
                            long="Long",stage1wt="wgt", stage2wt="Panelwgt",
                            str1prop=NA)
round(WLRDB_ests,4)

#      mu  trend SEtrend  sig2a sig2t sigat sig2b  sig2e   eta
#  4.1914 -0.0177  0.0492     0     0     0     0 0.0002   4


###########################
#  ---- C1. PWIGLS with type="Aonly".
PWIGLS_Aonly_ests <- TrendNPS_Cont(dat=SEKIANC, method="PWIGLS", type="Aonly",
                                   slope=FALSE, stratum=NA, Y="Y", lat="Lat",
                                   long="Long",stage1wt="wgt", 
                                   stage2wt="Panelwgt", str1prop=NA)
round(PWIGLS_Aonly_ests,4)

#      mu  trend SEtrend  sig2a sig2t sigat  sig2b  sig2e    eta
#  3.8812 0.0383   0.031 6.2650     0     0 0.0732 0.0167 3.6202

}
\author{
Leigh Ann Starcevich and Jason Mitchell of Western EcoSystems 
  Technology, Inc.
}
\references{
Asparouhov, T. (2006).  General multi-level modeling with 
  sampling weights. Communications in Statistics - Theory and Methods 35: 
  439-460.
  
  Little, J.A.R., and D.B. Rubin. 2002. Statistical analysis with missing 
  data, 2nd edition. John Wiley and Sons, Inc., New Jersey.
  
  Oh, H.L., and F.J. Scheuren. 1983. Weighting adjustment for unit 
  nonresponse. Pages 143-184 in W.G. Madow, I. Olkin, and D.B. Rubin, 
  editors. Incomplete data and in sample surveys. (Vol. 2). Academic Press, 
  New York.
  
  Pfeffermann, D., C.J. Skinner, D.J. Holmes, H. Goldstein, and J. Rasbash 
  (1998).  Weighting for unequal selection probabilities in multilevel 
  models. Journal of the Royal Statistical Society, Series B 60(1): 23-40.
  
  Piepho, H.P., & Ogutu, J.O. (2002).  A simple mixed model for trend 
  analysis in wildlife populations.  Journal of Agricultural, Biological, and
  Environmental Statistics, 7(3), 350-360.
  
  L.A. Starcevich, G. DiDonato, T. McDonald, and J. Mitchell. 2016. A GRTS 
  user's manual for the SDrawNPS package: A graphical user interface for 
  generalized random tessellation stratified (GRTS) sampling and estimation. 
  Natural Resource Report NPS/PWRO/NRR-2016/1233. National Park Service, Fort
  Collins, Colorado.
  
  L.A. Starcevich, T. McDonald, A. Chung-MacCoubrey, A. Heard, and J. Nesmith
  (2017). Trend Estimation for Complex Survey Designs. Natural Resource 
  Report NPS/xxxx/NRR-2017/xxxx. National Park Service, Fort Collins, 
  Colorado.
}
\seealso{
\code{lme4}, \code{lmerTest}, \code{spsurvey}
}

