% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearizationVar.r
\name{LinearizationVar}
\alias{LinearizationVar}
\title{Calculate linearization variance of trend for the PWIGLS trend approach.}
\usage{
LinearizationVar(Site, wij, xij, eij, varYij)
}
\arguments{
\item{Site}{vector of Site names.}

\item{wij}{vector of inclusion weights = Site-level design weight * Year-level panel weight.}

\item{xij}{vector of shifted year variables (WYear) for trend estimation.}

\item{eij}{vector of residuals from PWIGLS trend model.}

\item{varYij}{vector of total variance estimates for yij, estimated from PWIGLS variance 
components estimates.}
}
\value{
Scalar variance estimate of the trend regression coefficient.
}
\description{
Calculate the linearization variance of the trend coefficient with a Taylor
series approximation. This is an internal function called from PWIGLS_ALL.
}
\details{
The linearization variance is based on a Taylor series approximation. See 
Skinner (1989, p. 82-83) for more information.
}
\examples{
\dontrun{
#  ---- Read example data set.
TrendVar = LinearizationVar(Site,wij,xij,eij,varYij)
}


}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
\references{
Skinner, C. J., D. Holt, and T. M. F. Smith. 1989. Analysis of Complex Surveys. 
New York: Wiley.
}
\seealso{
\code{TrendNPS_Cont}, \code{LinearizationVar_StRS}
}

