% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWIGLS_ALL.r
\name{PWIGLS_ALL}
\alias{PWIGLS_ALL}
\title{Probability-weighted iterative generalized least squares (PWIGLS).}
\usage{
PWIGLS_ALL(Z, dat, stage1wt, stage2wt, type, stratum, slope)
}
\arguments{
\item{Z}{Random-effects design matrix from the unweighted (PO) model.}

\item{dat}{Data frame containing columns at least for \code{Site}, 
\code{WYear}, \code{Year}, and the continuous outcome of interest \code{Y}.
See section "Data frame \code{dat}" below.}

\item{stage1wt}{Design weights from the original sample draw without 
accounting for temporal revisit designs.}

\item{stage2wt}{Panel inclusion weights for each site each year.}

\item{type}{Scaling type when \code{method="PWIGLS"}. Valid values include 
\code{"Aonly"}, \code{"A"}, \code{"AI"}, \code{"B"}, \code{"BI"} 
\code{"C"}. See section "Options for variable \code{type}" below.}

\item{stratum}{Text string identifying an optional two-level stratification 
factor in \code{dat}. Use stratum = NA to indicate no stratification used.}

\item{slope}{Logical value indicating inclusion of a random site-level slope
effect in the variance components structure in addition to the Site- and Year-
level random intercept terms. Default = TRUE.}
}
\value{
Returns a vector of regression coefficient estimates for the trend model.
}
\description{
Calculate trend from data collected with complex survey designs by 
incorporating weights with a linear mixed model. This is an internal function 
called from TrendNPS_Cont.
}
\details{
Calculates the probability-weighted iterative generalized least squares
(PWIGLS) trend model (Pfeffermann et al. 1998; Asparouhov 2006).
}
\section{Options for variable \code{type}}{

  
  Selection of \code{method="PWIGLS"} requires further specification of 
  argument \code{type}.  Valid options include
  
  \tabular{ll}{
  
  \code{"Aonly"} \tab probability weighting but no scaling at either stage \cr
  \code{"A"}     \tab scaling for panel weights with the mean site-level design weight  \cr
  \code{"AI"}    \tab scaling for panel weights with the mean site-level design weight, but no site-level scaling  \cr
  \code{"B"}     \tab scaling for panel weights with the effective mean site-level design weight  \cr
  \code{"BI"}    \tab scaling for panel weights with the effective mean site-level design weight, but no site-level scaling  \cr
  \code{"C"}     \tab scaling only at the year level with the inverse of the average year-level weight \cr
  
  }
}

\examples{
\dontrun{
#  ---- Read example data set.
	fit<-PWIGLS_ALL(Z=getME(fit_PO,"Z"),dat=dat,stage1wt=stage1wt,stage2wt=stage2wt,type=type,
      stratum=stratum,slope=slope)
}


}
\references{
Asparouhov, T. (2006).  General multi-level modeling with 
  sampling weights. Communications in Statistics - Theory and Methods 35: 
  439-460.

  Pfeffermann, D., C.J. Skinner, D.J. Holmes, H. Goldstein, and J. Rasbash 
  (1998).  Weighting for unequal selection probabilities in multilevel 
  models. Journal of the Royal Statistical Society, Series B 60(1): 23-40.
}
\seealso{
\code{TrendNPS_Cont}, \code{LinearizationVar}
}
\author{
Leigh Ann Starcevich of Western EcoSystems Technology, Inc.
}
