% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendNPS.r
\docType{package}
\name{TrendNPS}
\alias{TrendNPS}
\alias{TrendNPS-package}
\title{TrendNPS: A package for dealing with trending stuff for the NPS.}
\description{
The TrendNPS package does trend stuff.  It's pretty cool.
}
\section{TrendNPS functions}{

TrendNPS_Cont
LinearizationVar_StRS
LinearizationVar
PWIGLS_ALL
}

