% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendNPS_Cont.r
\name{TrendNPS_Cont}
\alias{TrendNPS_Cont}
\title{Trend for Continuous Outcomes from Complex Survey Designs}
\usage{
TrendNPS_Cont(alpha, dat, method, slope = TRUE, type = NA, stratum = NA,
  Y, lat = NA, long = NA, stage1wt = NA, stage2wt = NA, str1prop = NA,
  nbhd = TRUE)
}
\arguments{
\item{alpha}{The probability of a Type I error.}

\item{dat}{Data frame containing columns at least for \code{Site}, 
\code{WYear}, \code{Year}, and the continuous outcome of interest \code{Y}.
See section "Data frame \code{dat}" below.}

\item{method}{Method for trend estimation entered as a string. Valid values 
include \code{"PO"} for the Piepho and Ogutu (2002) unreplicated linear 
mixed model, \code{"SLRDB"} for simple linear regression of design-based 
estimates, \code{"WLRDB"} for weighted linear regression of design-based 
estimates, or \code{"PWIGLS"} for probability-weighted iterative 
generalized least squares (Pfeffermann et al. 1998, Asparouhov 2002).}

\item{slope}{Logical value indicating inclusion of a random site-level slope
effect in the variance components structure used for the \code{"PO"} and 
\code{"PWIGLS"} trend methods.  Site- and year-level random intercept terms
included as default.}

\item{type}{Scaling type when \code{method="PWIGLS"}. Valid values include 
\code{"Aonly"}, \code{"A"}, \code{"AI"}, \code{"B"}, \code{"BI"} 
\code{"C"}. See section "Options for variable \code{type}" below.}

\item{stratum}{Text string identifying an optional two-level stratification 
factor in \code{dat}.}

\item{Y}{Text string indicating the outcome variable in \code{dat}.}

\item{lat}{Site latitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{long}{Site longitudes using an equal-area projection, e.g., UTM or 
Albers.}

\item{stage1wt}{Design weights from the original sample draw without 
accounting for temporal revisit designs.}

\item{stage2wt}{Panel inclusion weights for each site each year.}

\item{str1prop}{Proportion of the first stratum in the population in a 
two-level stratification variable.}

\item{nbhd}{Logical indictor for using the neighborhood variance estimator 
when the \code{method="WLRDB"}.  If \code{FALSE}, \code{"WLRDB"} calculates
the design-based mean estimate's standard error assuming independent random
sampling.}
}
\value{
A list containing three or four elements, depending on the specified trend method. 
  The first element of the list contains a data frame containing the estimated intercept,  
  trend estimate, trend standard error, estimated variance components, and degrees of freedom  
  used for trend testing and confidence intervals construction.
  
  The contents of the first list element, ModelEstimates, are:
  
  \tabular{rl}{
  
  \code{mu}      \tab	Estimated intercept from the trend model/ \cr
  \code{trend}   \tab	Estimated trend of the logged outcome \code{Y} from the trend model.  \cr
  \code{SEtrend} \tab	Estimated standard error of the trend estimate for the logged outcome \code{Y}. \cr
  \code{sig2a}   \tab	Estimated site-to-site variation. \cr
  \code{sig2b}   \tab	Estimated year-to-year variation. \cr
  \code{sig2t}   \tab	Estimated variation among site-level slopes. \cr
  \code{sig2e}   \tab	Estimated residual variation. \cr
  \code{eta}     \tab	Degrees of freedom used to test for trend and calculate confidence intervals. \cr
  
  }
  
  The contents of the second list element, TrendTest, include the trend coefficient on the logged scale, 
  the trend standard error, the t-statistic, degrees of freedom, and p-value for a two-sided test of no trend.
  
  The contents of the third list element, TrendCI, include the back-transformed trend on the original scale 
  calculated as an annual proportional change and the lower and upper (1-\code{alpha})*100\%-confidence interval bounds. 
  
  A fourth list element, DBests, contains the annual design-based estimates and is returned only when the 
  SLRDB and WLRDB trend approaches are used.  The design-based estimates may be used to assess the assumption
  of independence for the linear regression models.  The output of the DBests list element include:
  
  \tabular{rl}{
  
  \code{Year}      \tab	Year of survey. \cr
  \code{Est.Mean}   \tab	Design-based estimate of the mean of the outcome of interest. \cr
  \code{SE} \tab	   Estimated standard error of design-based mean estimate via neighborhood 
  variace estimator if \code{nbhd = TRUE}. \cr
  \code{WYear}   \tab	WYear variable. \cr
  \code{Resid}     \tab	Residuals from the linear model fit. \cr
  
  }
}
\description{
\code{TrendNPS_Cont} fits trend models of continuous outcomes 
  for four approaches:
  
  \enumerate{ \item the unreplicated linear mixed model as specified by 
  Piepho and Ogutu (2002) that does not incorporate design weights ("PO");
  
  \item simple linear regression of annual design-based estimates ("SLRDB");
  
  \item weighted linear regression of annual design-based estimates (WLRDB);
  
  \item and probability-weighted iterative generalized least squares of a 
  linear mixed model (PWIGLS) with 6 types of scaling.
  
  } Fixed effects structure includes a year term for trend estimation and an 
  optional two-level stratification factor. The function assumes random site 
  and year intercepts, and an optional random site-level slope effect may be 
  included.
}
\details{
For long-term monitoring, \code{TrendNPS_Cont} assumes single annual sampling 
  occasions per unique record of \code{WYear} and \code{Site} in data frame \code{dat}.  
  
  Sampling (\code{stage1wt}) and temporal-revisit (\code{stage2wt}) design 
  weights must be adjusted for any missing data or frame error prior to trend
  analysis. See Oh and Scheuren (1983), Little and Rubin (2002), and 
  Starcevich et al. (2016).
  
  For background on the unweighted linear mixed model, see Piepho & Ogutu
  (2002).  For information on the probability-weighted generalized least
  squares approach, see Pfeffermann et al. (1998), Asparouhov (2006), and
  Starcevich et al. (2017).
}
\section{Method requirements}{

  
  Spatially balanced samples utilizing \code{method="SLRDB"} or 
  \code{method="WLRDB"} require specification of \code{lat} and \code{long} 
  on setting the neighborhood variance estimator \code{nbhd=TRUE}.  If 
  spatial balance cannot be assumed, set option \code{nbhd=FALSE} for 
  non-spatially balanced designs so that the \code{"SLRDB"} and 
  \code{"WLRDB"} methods assume independent random sampling for standard 
  errors.
  
  The stratification variable \code{stratum} is required for all methods when
  design strata are used.  Note that \code{method="PWIGLS"} further requires
  a value for \code{type}.  See section "Options for variable \code{type}"
  below.
  
  Variables \code{stage1wt} and \code{stage2wt} must be specified for all 
  methods except the unweighted \code{method="PO"}.
}

\section{Data frame \code{dat}}{

  
  Data frame \code{dat} requires at least variables \code{Site}, 
  \code{WYear}, \code{Year}, and \code{Y}.  Variable \code{Site} defines the 
  sampling unit.
  
  Underlying statistical functions called by \code{TrendNPS_Cont} require two
  separate temporal variables: \code{Year} is a factor and \code{WYear} is a 
  scalar-year value.  \code{WYear} is shifted so that the year for which \code{Y}
  is least variable is centered on 0. See Piepho & Ogutu (2002) for more 
  information.
  
  Ultimately, \code{WYear} represents time in the fixed-effects portion of the  
  trend model. The regression coefficient associated with \code{WYear} provides
  the basis for trend estimation and testing. 
  
  Variable \code{Year} enters the model as a random effect and provides 
  the basis for estimation of random year-to-year variation among survey 
  occasions.
  
  Variable \code{Y} must contain only positive non-zero values.  Trend analyses 
  approaches will implement a logarithmic transformation for trend modeling of 
  variable \code{Y}, but the untransformed variable \code{Y} should be provided 
  as an function call.
}

\section{Options for variable \code{type}}{

  
  Selection of \code{method="PWIGLS"} requires further specification of 
  argument \code{type}.  Valid options include
  
  \tabular{rl}{
  
  \code{"Aonly"} \tab Probability weighting but no scaling at either stage. \cr
  \code{"A"}     \tab Scaling for panel weights with the mean site-level design weight.  \cr
  \code{"AI"}    \tab Scaling for panel weights with the mean site-level design weight, but no site-level scaling.  \cr
  \code{"B"}     \tab Scaling for panel weights with the effective mean site-level design weight.  \cr
  \code{"BI"}    \tab Scaling for panel weights with the effective mean site-level design weight, but no site-level scaling.  \cr
  \code{"C"}     \tab Scaling only at the year level with the inverse of the average year-level weight. \cr
  
  }
}

\examples{

#  ---- Read example data set.
data(LakeDataRep)

#  ---- Load dependent packages. 
pkgList <- c("lme4","lmerTest","spsurvey")
inst <- pkgList \%in\% installed.packages()
if (length(pkgList[!inst]) > 0) install.packages(pkgList[!inst])
lapply(pkgList, library, character.only = TRUE)

###########################
# Example 1: Trend analysis with the PO approach: stratification and full random 
# effects model.
str1prop <- 808.6936/(808.6936+362.8214)
PO_ests <- TrendNPS_Cont(alpha=0.1,dat=LakeDataRep, method="PO", slope=TRUE, 
     stratum="Park", Y="Cl", str1prop=str1prop)
PO_ests
# $ModelEstimates
#         mu       trend    SEtrend     sig2a       sig2t       sigat      sig2b
#   1.269532 -0.07196304 0.05624014 0.2846501 0.009663692 -0.05208829 0.04079863
#      sig2e      eta
# 0.08133577 24.94572

# $TrendTest
#        trend    SEtrend    t-stat      eta    pvalue
#  -0.07196304 0.05624014 -1.279567 24.94572 0.2124709

# $TrendCI
#  Annual Pct Change     CI low   CI high
#        -0.06943471 -0.1546776 0.0244041

###########################
# Example 2: Trend analysis with the WLRDB approach: stratification and full random effects model.
str1prop <- 808.6936/(808.6936+362.8214)
WLRDB_ests <- TrendNPS_Cont(alpha=0.1,dat=LakeDataRep, method="WLRDB",  
         slope=TRUE, stratum="Park", Y="Cl", lat="ycoord",
         long="xcoord",stage1wt="AdjWgt", stage2wt="PanelWt",
         str1prop=str1prop)
WLRDB_ests
# $ModelEstimates
#         mu      trend    SEtrend sig2a sig2t sigat sig2b    sig2e eta
#  0.4218735 0.06037018 0.07802896     0     0     0     0 3.402272   4

# $TrendTest
#       trend    SEtrend    t-stat eta    pvalue
#  0.06037018 0.07802896 0.7736894   4 0.4822974

# $TrendCI
#  Annual Pct Change     CI low   CI high
#         0.06222969 -0.1005534 0.2544735

#$DBests
#  Year Est.Mean         SE WYear        Resid
#  2008 4.505724 0.55610529     0  1.083475074
#  2009 2.162006 0.20158290     1  0.288793000
#  2010 1.712219 0.03578556     2 -0.004823561
#  2011 1.592197 0.07584161     3 -0.137869385
#  2012 2.611047 0.14932238     4  0.296397135
#  2013 2.348186 0.18412120     5  0.129918636


###########################
# Example 3: Trend analysis with the PWIGLS method with type="Aonly".
str1prop <- 808.6936/(808.6936+362.8214)
PWIGLS_Aonly_ests <- TrendNPS_Cont(alpha=0.1,dat=LakeDataRep, method="PWIGLS", 
         type="Aonly", slope=TRUE, stratum="Park", Y="Cl", lat="ycoord",
         long="xcoord", stage1wt="AdjWgt", stage2wt="PanelWt",
         str1prop=str1prop)
PWIGLS_Aonly_ests

# $ModelEstimates
#       mu       trend    SEtrend    sig2a     sig2t     sigat     sig2b
# 1.441448 -0.08109772 0.09649006 10.66269 0.3102913 -1.818939 0.4271289
#      sig2e      eta
# 0.09145612 24.94572

# $TrendTest
#       trend    SEtrend     t-stat      eta    pvalue
# -0.08109772 0.09649006 -0.8404774 24.94572 0.4086247

# $TrendCI
#  Annual Pct Change     CI low    CI high
#        -0.07789642 -0.2180231 0.08734037

}
\references{
Asparouhov, T. (2006).  General multi-level modeling with 
  sampling weights. Communications in Statistics - Theory and Methods 35: 
  439-460.
  
  Little, J. A. R., and D. B. Rubin. (2002). Statistical analysis with missing 
  data, 2nd edition. John Wiley and Sons, Inc., New Jersey.
  
  Oh, H. L., and F. J. Scheuren. (1983). Weighting adjustment for unit 
  nonresponse. Pages 143-184 in W.G. Madow, I. Olkin, and D.B. Rubin, 
  editors. Incomplete data and in sample surveys. (Vol. 2). Academic Press, 
  New York.
  
  Pfeffermann, D., C. J. Skinner, D. J. Holmes, H. Goldstein, and J. Rasbash 
  (1998).  Weighting for unequal selection probabilities in multilevel 
  models. Journal of the Royal Statistical Society, Series B 60(1): 23-40.
  
  Piepho, H. P., & Ogutu, J. O. (2002).  A simple mixed model for trend 
  analysis in wildlife populations.  Journal of Agricultural, Biological, and
  Environmental Statistics, 7(3), 350-360.
  
  L. A. Starcevich, G. DiDonato, T. McDonald, and J. Mitchell. (2016). A GRTS 
  user's manual for the SDrawNPS package: A graphical user interface for 
  generalized random tessellation stratified (GRTS) sampling and estimation. 
  Natural Resource Report NPS/PWRO/NRR-2016/1233. National Park Service, Fort
  Collins, Colorado.
  
  L. A. Starcevich, T. McDonald, A. Chung-MacCoubrey, A. Heard, and J. Nesmith.
  (2017). Trend Estimation for Complex Survey Designs. Natural Resource 
  Report NPS/xxxx/NRR-2017/xxxx. National Park Service, Fort Collins, 
  Colorado.
}
\seealso{
\code{lme4}, \code{lmerTest}, \code{spsurvey}
}
\author{
Leigh Ann Starcevich and Jason Mitchell of Western EcoSystems 
  Technology, Inc.
}
